add private methods without requests and patch, add findById method

+++ ewm-service/src/main/java/ru/rubik/ewmservice/event/requests/EventCreateRequest.java
@@ -17,7 +17,7 @@ public class EventCreateRequest {
     private String annotation;
     private LocalDateTime eventDate;
     private Boolean paid;
-    private Integer category;
+    private Long category;
     private Integer participantLimit;
     private Boolean requestModeration;
     //todo -> location
